fix(devcontainer): setup

+++ justfile
@@ -5,7 +5,9 @@ init:
     #!/bin/zsh
     set -euxo pipefail
     git config --global --add safe.directory '*'
-    cp /tmp/.zshrc /home/vscode/.zshrc
+    git config --global core.autocrlf input
+    git config --global core.eol lf
+    if [ -f /tmp/.zshrc ]; then cp /tmp/.zshrc /home/vscode/.zshrc; fi
     rm -rf .venv
     uv venv
     uv sync --all-groups
